backend/pyproject.toml
[tool.poetry]
name = "agentforge-backend"
version = "1.0.0"
description = "Self-Healing Multi-Agent Coding Assistant - Backend"
authors = ["Your Name <you@email.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
google-generativeai = "^0.8.0"
langchain = "^0.3.0"
langgraph = "^0.2.0"
langsmith = "^0.1.0"
langchain-google-genai = "^2.0.0"
neo4j = "^5.23.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
